pyproject.toml
[tool.poetry]
name = "aws-cdk-iac"
version = "0.1.0"
description = ""
authors = ["sanggubot <sanggubot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.118.0"
boto3 = "^1.34.15"
python-dotenv = "^1.0.0"
boto3-stubs = "^1.34.17"
mypy = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
